Add files via upload
A Database Solution for Brewed Coffee Shop
This project presents a complete relational database design for Brewed Coffee Shop, a cozy beverage shop offering a variety of drinks and desserts. I led the design and modeling of a normalized database system to support critical business operations including:

     -Order and payment processing

     -Inventory and ingredient tracking

     -Menu management

     -Customer loyalty programs

The project includes a conceptual EER diagram, logical schema with table structures and relationships, and defined domain constraints. It demonstrates practical database design skills, use of specialization/generalization, and an understanding of many-to-many relationships and normalization principles. Included are images of the implemented database running on a MySQL server. These images include the creation of every table in the schema, as well as 10 queries made to test the schema—alongside the results of said queries.